+++ src/routes/(app)/communications/whatsapp/[thread_id]/+page.svelte
@@ -97,42 +97,8 @@
 	import * as m from '$lib/paraglide/messages';
 	import { goto } from '$app/navigation';
 
-	import { countTextTemplatePlaceholders } from '$lib/comps/forms/whatsapp/messages/builder/actions/components';
-
-	function hasUnfilledPlaceholders() {
-		if (!template || templateMessage.type !== 'template') return false;
-
-		// Check header
-		const headerComponent = template.message.components.find(
-			(comp) => comp.type === 'HEADER' && comp.format === 'TEXT'
-		);
-		if (headerComponent && 'text' in headerComponent) {
-			const headerPlaceholders = countTextTemplatePlaceholders(headerComponent.text);
-			const headerParams = components.find((comp) => comp.type === 'header')?.parameters || [];
-			if (headerPlaceholders > headerParams.length) return true;
-
-			for (const param of headerParams) {
-				if (param.type === 'text' && /{{[0-9]+}}/.test(param.text)) return true;
-			}
-		}
-
-		// Check body
-		const bodyComponent = template.message.components.find((comp) => comp.type === 'BODY');
-		if (bodyComponent && 'text' in bodyComponent) {
-			const bodyPlaceholders = countTextTemplatePlaceholders(bodyComponent.text);
-			const bodyParams = components.find((comp) => comp.type === 'body')?.parameters || [];
-			if (bodyPlaceholders > bodyParams.length) return true;
-
-			for (const param of bodyParams) {
-				if (param.type === 'text' && /{{[0-9]+}}/.test(param.text)) return true;
-			}
-		}
-
-		return false;
-	}
-
 	async function handleSendClick() {
-		if (hasUnfilledPlaceholders()) {
+		if (componentActions.hasUnfilledPlaceholders(template, components, templateMessage)) {
 			$flash = { type: 'error', message: 'Please fill in all placeholders before sending' };
 			return;
 		}
